requirements.txt
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
decorator==4.4.2
idna==3.6
imageio==2.34.0
imageio-ffmpeg==0.4.9
joblib==1.3.2
Markdown==3.4.3
moviepy==1.0.3
numpy==1.26.4
Pillow==9.5.0
praw==7.7.1
prawcore==2.4.0
profanity-check==1.0.3
proglog==0.1.10
pydub==0.25.1
pytube==15.0.0
requests==2.31.0
scikit-learn==1.4.1.post1
scipy==1.12.0
soupsieve==2.5
threadpoolctl==3.3.0
tqdm==4.66.2
update-checker==0.18.0
urllib3==2.2.1
websocket-client==1.7.0
youtube-search==2.1.2